ecoshop-main/leaderboard/requirements.txt
flask
supabase
python-dotenv
flask_cors